rebase post #1240
