+++ packages/legend-application/src/application/LegendApplication.tsx
@@ -33,10 +33,7 @@ import {
 } from '@finos/legend-shared';
 import { APPLICATION_EVENT } from '../stores/ApplicationEvent.js';
 import { configureComponents } from '@finos/legend-art';
-import type {
-  GraphPluginManager,
-  PackageableElement,
-} from '@finos/legend-graph';
+import type { GraphPluginManager } from '@finos/legend-graph';
 import type { LegendApplicationPluginManager } from './LegendApplicationPluginManager.js';
 import { setupPureLanguageService } from '../stores/PureLanguageSupport.js';
 
@@ -271,21 +268,3 @@ export abstract class LegendApplication {
     }
   }
 }
-
-export const packageableElementFormatOptionLabel = (option: {
-  label: string;
-  value: PackageableElement;
-  darkMode?: boolean;
-}): React.ReactNode => {
-  const stylePrefix = option.darkMode
-    ? 'packageable-element-format-option-label--dark'
-    : 'packageable-element-format-option-label';
-  return (
-    <div className={stylePrefix}>
-      <div className={`${stylePrefix}__name`}>{option.label}</div>
-      {option.value.package && (
-        <div className={`${stylePrefix}__tag`}>{`${option.value.path}`}</div>
-      )}
-    </div>
-  );
-};
