+++ packages/legend-application/src/index.ts
@@ -48,6 +48,7 @@ export * from './components/shared/TextInputEditor.js';
 export * from './components/shared/LambdaEditor.js';
 export * from './components/shared/BasicValueSpecificationEditor.js';
 export * from './components/shared/LambdaParameterValuesEditor.js';
+export * from './components/shared/PackageableElementOption.js';
 export * from './components/shared/execution-plan-viewer/ExecutionPlanViewer.js';
 
 export { LambdaEditorState } from './stores/shared/LambdaEditorState.js';
