+++ packages/legend-application/style/components/shared/_packageable-element-format-option.scss
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 @use 'mixins' as *;
 
 .packageable-element-format-option-label {
